Rename BuildApkTemporary.
